Move BlockType enum to Blocks namespace
