+++ app/Enums/Blocks/BlockType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Enums;
+namespace App\Enums\Blocks;
 
 enum BlockType: string
 {
