+++ app/Http/Resources/Views/Blocks/StackBlockResource.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Resources\Views\Blocks;
 
-use App\Enums\BlockType;
+use App\Enums\Blocks\BlockType;
 use App\Http\Resources\Formatters\LaravelVersionFormatterResource;
 use App\Http\Resources\Formatters\PhpVersionFormatterResource;
 use App\Interfaces\Resources\Items\ConstantItemInterface;
