+++ database/factories/BlockFactory.php
@@ -2,7 +2,7 @@
 
 namespace Database\Factories;
 
-use App\Enums\BlockType;
+use App\Enums\Blocks\BlockType;
 use App\Models\Block;
 use App\Models\Page;
 use Illuminate\Database\Eloquent\Factories\Factory;
