+++ tests/Datasets/BlockTypes.php
@@ -1,5 +1,5 @@
 <?php
 
-use App\Enums\BlockType;
+use App\Enums\Blocks\BlockType;
 
 dataset('block-types', fn () => BlockType::cases());
